backend-app/requirements.txt
fastapi
uvicorn
google-generativeai
pandas
gspread
oauth2client
geopy
python-dotenv

